requirements.txt
asgiref==3.8.1
contourpy==1.2.1
cycler==0.12.1
Django==4.1.13
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.6.1
fonttools==4.53.1
kiwisolver==1.4.5
matplotlib==3.9.2
numpy==2.1.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
PyJWT==2.9.0
pymongo==4.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.2.4
tzdata==2024.1
